% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Based on protein expressions, assign each cell to their category. This is achieved through
a predefined rule.}
\usage{
annotate(medexpression)
}
\arguments{
\item{medexpression}{The expression profile of each cluster identified by spade}
}
\value{
A vector of annotations for each cell
}
\description{
Based on protein expressions, assign each cell to their category. This is achieved through
a predefined rule.
}
