% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{gating}
\alias{gating}
\title{Apply gating to a set of raw fcs files in a specified folder.
This function applies filter to a number of fcs files in the folder and combine them into a single
fcs file and another text file (a table) that matches each cell event to the original patient file.
The two files are named by default as "output".}
\usage{
gating(folder, output_name = "output")
}
\arguments{
\item{folder}{The folder where the function probes through all the fcs files.}

\item{output_name}{The output file name for the fcs file.}
}
\value{
A vector with the two output filenames
}
\description{
Apply gating to a set of raw fcs files in a specified folder.
This function applies filter to a number of fcs files in the folder and combine them into a single
fcs file and another text file (a table) that matches each cell event to the original patient file.
The two files are named by default as "output".
}
