% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_summary.R
\name{QC_summary}
\alias{QC_summary}
\title{Quality control of a folder of raw fcs files. Combine them, apply filter and offer a summary of cell counts for each patient.}
\usage{
QC_summary(folder, spade_save = FALSE)
}
\arguments{
\item{folder}{A folder where the fcs files live in}

\item{spade_save}{whether to save the intermediate files of spade analysis. By default false.}
}
\value{
check out function 'gating' and 'spade_analysis'.
}
\description{
Quality control of a folder of raw fcs files. Combine them, apply filter and offer a summary of cell counts for each patient.
}
\examples{
QC_summary(folder = system.file("extdata", package = "CyTOFprocess"))


}
