% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{single_cell_filtering}
\alias{single_cell_filtering}
\title{Apply filtering on a single FCS flowframe}
\usage{
single_cell_filtering(cytoframe)
}
\arguments{
\item{cytoframe}{A flowframe object defined in package 'flowCore'.}
}
\value{
A data table object of a subset of the original flowframe expression matrix.
}
\description{
Apply filtering on a single FCS flowframe
}
