% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADE.R
\name{spade_analysis}
\alias{spade_analysis}
\title{Apply SPADE analysis on an fcs file combined with its metadata and offer summary of cell counts
for each patient. The metadata contains information about the patient each cell belongs to.}
\usage{
spade_analysis(filename, intermediate = FALSE, patient_id,
  output = "SPADE_out")
}
\arguments{
\item{filename}{The input fcs file name}

\item{intermediate}{Whether to save the intermediate files generate through the SPADE.driver function in the spade package
developed at Garry Nolan's lab}

\item{patient_id}{The input patient ID txt file corresponding to the fcs file}

\item{output}{The output folder that stores the spade analysis results}
}
\value{
A tibble dataframe that summarizes the number of cells in each patient
}
\description{
Apply SPADE analysis on an fcs file combined with its metadata and offer summary of cell counts
for each patient. The metadata contains information about the patient each cell belongs to.
}
